+++ R/modspec.R
@@ -422,7 +422,7 @@ specMATRIX <- function(x,
                        oclass,type, annotated = FALSE,
                        env, pos=1,
                        name="...", prefix="", labels=NULL,
-                       object=NULL,unlinked=FALSE,...) {
+                       object=NULL, unlinked=FALSE,...) {
   
   if(is.null(object)) check_block_data(x,env$ENV,pos)
   
@@ -436,12 +436,17 @@ specMATRIX <- function(x,
       unlinked <- FALSE
       novalue <- FALSE
     } else {
-      stop("Ambigious or mixed annotations in ",paste0("$",toupper(type)),call.=FALSE) 
+      stop(
+        "Ambigious or mixed annotations in ",paste0("$",toupper(type)),
+        call.=FALSE
+      ) 
     }
     
-    l <- parse_annot(x[anl],name_value=FALSE,
-                     block=toupper(type),
-                     envir=env$ENV,novalue=novalue)
+    l <- parse_annot(
+      x[anl],name_value=FALSE,
+      block=toupper(type),
+      envir=env$ENV,novalue=novalue
+    )
     
     if(unlinked) {
       l[["v"]] <- as.numeric(cvec_cs(x[!anl])) 
@@ -450,6 +455,13 @@ specMATRIX <- function(x,
     d <- modMATRIX(l[["v"]],context=oclass,...)
     labels <- l[["an"]][["name"]]
     env[["annot"]][[pos]] <- l[["an"]]
+    
+    if(unlinked & nrow(d) != length(labels)) {
+      stop(
+        "Annotated matrix in unlinked configuration is misspecified", 
+        call. = FALSE
+      )
+    }
     
   } else {
     if(any(anl)) x <- x[!anl]
@@ -479,19 +491,22 @@ specMATRIX <- function(x,
 
 ##' @export
 handle_spec_block.specOMEGA <- function(x,...) {
-  scrape_and_call(x,
-                  pass="specMATRIX",
-                  def=list(oclass="omegalist",type="omega"),
-                  narrow=FALSE,...)
+  scrape_and_call(
+    x,
+    pass="specMATRIX",
+    def=list(oclass="omegalist",type="omega"),
+    narrow=FALSE,...
+  )
 }
 
 ##' @export
 handle_spec_block.specSIGMA <- function(x,...) {
-  scrape_and_call(x,
-                  pass="specMATRIX",
-                  def=list(oclass="sigmalist",type="sigma"),
-                  narrow=FALSE,...)
-  
+  scrape_and_call(
+    x,
+    pass="specMATRIX",
+    def=list(oclass="sigmalist",type="sigma"),
+    narrow=FALSE,...
+  )
 }
 
 eval_ENV_block <- function(x,where,envir=new.env(),...) {